backend/pyproject.toml
[project]
name = "rescueforge-backend"
version = "0.5.0"
description = "AI-powered fire department floor plan generator"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "ezdxf>=1.4.0",
    "shapely>=2.0.0",
    "svgwrite>=1.4.3",
    "weasyprint>=63.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "pillow>=11.0.0",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "websockets>=14.0",
    "python-docx>=1.1.0",
    "networkx>=3.0",
    "scipy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
    "cairosvg>=2.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
